move init server
